add roles and security in the traceability

+++ backend/routes/traceability_routes.py
@@ -1,14 +1,18 @@
 from flask import Blueprint, request, jsonify
-from datetime import datetime
+from flask_jwt_extended import get_jwt_identity
+from datetime import datetime, timedelta
 from models.db import db
 from models.traceability import Traceability
 from models.patient import Patient
 from models.medical_product import MedicalProduct
+from utils.role_required import role_required
+from enums import RoleEnum
 
 traceabilities_bp = Blueprint('traceabilities', __name__, url_prefix='/api/traceabilities')
 
 
 @traceabilities_bp.route('/', methods=['POST'])
+@role_required(RoleEnum.NURSE)
 def create_traceability():
     try:
         if not request.is_json:
@@ -23,9 +27,12 @@ def create_traceability():
         if not MedicalProduct.query.get(data.get('id_product')):
             return jsonify({"msg": "Product not found"}), 404
 
+        id_nurse_logueado=int(get_jwt_identity())
+
         new_traceability = Traceability(
             id_patient=data.get('id_patient'),
             id_product=data.get('id_product'),
+            id_nurse=id_nurse_logueado,
             date_of_use=datetime.fromisoformat(data.get('date_of_use')) if data.get('date_of_use') else datetime.utcnow()
         )
         db.session.add(new_traceability)
@@ -39,6 +46,7 @@ def create_traceability():
 
 
 @traceabilities_bp.route('/', methods=['GET'])
+@role_required(RoleEnum.NURSE, RoleEnum.DOCTOR)
 def get_traceabilities():
     try:
         traceabilities = Traceability.query.all()
@@ -48,6 +56,7 @@ def get_traceabilities():
 
 
 @traceabilities_bp.route('/<int:traceability_id>', methods=['GET'])
+@role_required(RoleEnum.NURSE, RoleEnum.DOCTOR)
 def get_traceability(traceability_id):
     try:
         traceability = Traceability.query.get(traceability_id)
@@ -59,6 +68,7 @@ def get_traceability(traceability_id):
 
 
 @traceabilities_bp.route('/patient/<int:patient_id>', methods=['GET'])
+@role_required(RoleEnum.NURSE, RoleEnum.DOCTOR)
 def get_traceabilities_by_patient(patient_id):
     try:
         traceabilities = Traceability.query.filter_by(id_patient=patient_id).all()
@@ -68,14 +78,24 @@ def get_traceabilities_by_patient(patient_id):
 
 
 @traceabilities_bp.route('/<int:traceability_id>', methods=['PUT'])
+@role_required(RoleEnum.NURSE)
 def update_traceability(traceability_id):
     try:
         traceability = Traceability.query.get(traceability_id)
         if not traceability:
             return jsonify({"msg": "Traceability not found"}), 404
 
+        id_nurse_logueado=int(get_jwt_identity())
+
+        if traceability.id_nurse!= id_nurse_logueado:
+            return jsonify({"msg":"Only the nurse who recorded this traceability can edit it."}),403
+
+        limite = traceability.created_at + timedelta(minutes=5)
+        if datetime.utcnow() > limite:
+            return jsonify({"msg": "No se puede editar una trazabilidad pasados los 5 minutos de su creación"}), 403
+
         if not request.is_json:
-            return jsonify({"msg": "Missing JSON in request"}), 400
+            return jsonify({"msg":"Missing JSON in request"}),400
 
         data = request.get_json()
 
@@ -96,11 +116,16 @@ def update_traceability(traceability_id):
 
 
 @traceabilities_bp.route('/<int:traceability_id>', methods=['DELETE'])
+@role_required(RoleEnum.NURSE)
 def delete_traceability(traceability_id):
     try:
         traceability = Traceability.query.get(traceability_id)
         if not traceability:
             return jsonify({"msg": "Traceability not found"}), 404
+
+        id_nurse_logueado=int(get_jwt_identity())
+        if traceability.id_nurse!=id_nurse_logueado:
+            return jsonify({"msg":"Only the nurse who recorded this traceability can edit it."}),403
 
         db.session.delete(traceability)
         db.session.commit()
